backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
deepface
requests
beautifulsoup4
numpy
tensorflow
tf-keras
opencv-python
pillow
